working version

+++ simulation/lbm.c
@@ -74,14 +74,14 @@ int main(int argc, char* argv[])
     speed_t* cells     = NULL;    /* grid containing fluid densities */
     speed_t* tmp_cells = NULL;    /* scratch space */
     int*     obstacles = NULL;    /* grid indicating which cells are blocked */
-    double*  av_vels   = NULL;    /* a record of the av. velocity computed for each timestep */
+    float*  av_vels   = NULL;    /* a record of the av. velocity computed for each timestep */
 
     int    ii;                    /*  generic counter */
     struct timeval timstr;        /* structure to hold elapsed time */
     struct rusage ru;             /* structure to hold CPU time--system and user */
-    double tic,toc;               /* floating point numbers to calculate elapsed wallclock time */
-    double usrtim;                /* floating point number to record elapsed user CPU time */
-    double systim;                /* floating point number to record elapsed system CPU time */
+    float tic,toc;               /* floating point numbers to calculate elapsed wallclock time */
+    float usrtim;                /* floating point number to record elapsed user CPU time */
+    float systim;                /* floating point number to record elapsed system CPU time */
 
     parse_args(argc, argv, &final_state_file, &av_vels_file, &param_file);
 
@@ -97,9 +97,9 @@ int main(int argc, char* argv[])
         av_vels[ii] = av_velocity(params, cells, obstacles);
 
         #ifdef DEBUG
-        printf("==timestep: %d==\n", ii);
-        printf("av velocity: %.12E\n", av_vels[ii]);
-        printf("tot density: %.12E\n", total_density(params, cells));
+            printf("==timestep: %d==\n", ii);
+            printf("av velocity: %.12E\n", av_vels[ii]);
+            printf("tot density: %.12E\n", total_density(params, cells));
         #endif
     }
 
@@ -124,16 +124,16 @@ int main(int argc, char* argv[])
 }
 
 void write_values(const char * final_state_file, const char * av_vels_file,
-    const param_t params, speed_t* cells, int* obstacles, double* av_vels)
+    const param_t params, speed_t* cells, int* obstacles, float* av_vels)
 {
     FILE* fp;                     /* file pointer */
     int ii,jj,kk;                 /* generic counters */
-    const double c_sq = 1.0/3.0;  /* sq. of speed of sound */
-    double local_density;         /* per grid cell sum of densities */
-    double pressure;              /* fluid pressure in grid cell */
-    double u_x;                   /* x-component of velocity in grid cell */
-    double u_y;                   /* y-component of velocity in grid cell */
-    double u;                     /* norm--root of summed squares--of u_x and u_y */
+    const float c_sq = 1.0/3.0;  /* sq. of speed of sound */
+    float local_density;         /* per grid cell sum of densities */
+    float pressure;              /* fluid pressure in grid cell */
+    float u_x;                   /* x-component of velocity in grid cell */
+    float u_y;                   /* y-component of velocity in grid cell */
+    float u;                     /* norm--root of summed squares--of u_x and u_y */
 
     fp = fopen(final_state_file, "w");
 
@@ -209,17 +209,17 @@ void write_values(const char * final_state_file, const char * av_vels_file,
     fclose(fp);
 }
 
-double calc_reynolds(const param_t params, speed_t* cells, int* obstacles)
+float calc_reynolds(const param_t params, speed_t* cells, int* obstacles)
 {
-    const double viscosity = 1.0 / 6.0 * (2.0 / params.omega - 1.0);
+    const float viscosity = 1.0 / 6.0 * (2.0 / params.omega - 1.0);
 
     return av_velocity(params,cells,obstacles) * params.reynolds_dim / viscosity;
 }
 
-double total_density(const param_t params, speed_t* cells)
+float total_density(const param_t params, speed_t* cells)
 {
     int ii,jj,kk;        /* generic counters */
-    double total = 0.0;  /* accumulator */
+    float total = 0.0;  /* accumulator */
 
     for (ii = 0; ii < params.ny; ii++)
     {
